Final renaming done
