+++ tooling-server-configuration/src/main/java/org/jboss/eap/qe/microprofile/tooling/server/log/LogChecker.java
@@ -1,24 +1,24 @@
-package org.jboss.eap.qe.microprofile.tooling.server.logwatch;
+package org.jboss.eap.qe.microprofile.tooling.server.log;
 
 import java.util.regex.Pattern;
 
 /**
- * Interface describing log watcher which checks log for a pattern.
+ * Interface describing utility which checks a log
  */
-public interface LogReader {
+public interface LogChecker {
 
     /**
-     * Perform search in log or its excerpt and find if a line that matches the pattern.
+     * Perform search in log or its excerpt and find if a line matches the pattern.
      * @param pattern a pattern which will be the log line matched against
      * @return true if log line matching pattern was found in log, false otherwise
      */
-    boolean wasLineLogged(final Pattern pattern);
+    boolean logMatches(final Pattern pattern);
 
     /**
      * Perform search in log or its excerpt and find if a line contains a sub string.
      * @param subString a sub string which will be used to perform search among lines
      * @return true if log line containing a sub string was found in log, false otherwise
      */
-    boolean wasLineLogged(final String subString);
+    boolean logContains(final String subString);
 
 }
