+++ tooling-server-configuration/src/main/java/org/jboss/eap/qe/microprofile/tooling/server/log/ModelNodeLogChecker.java
@@ -1,4 +1,4 @@
-package org.jboss.eap.qe.microprofile.tooling.server.logwatch;
+package org.jboss.eap.qe.microprofile.tooling.server.log;
 
 import org.jboss.dmr.ModelNode;
 import org.wildfly.extras.creaper.core.online.ModelNodeResult;
@@ -13,10 +13,10 @@ import java.util.function.Predicate;
 import java.util.regex.Pattern;
 
 /**
- * Log Watcher based on accessing messages through {@link org.jboss.dmr.ModelNode} returned from {@code :read-log-file}
+ * Log checker based on accessing messages through {@link org.jboss.dmr.ModelNode} returned from {@code :read-log-file}
  * operation.
  */
-public final class ModelNodeLogReader implements LogReader {
+public final class ModelNodeLogChecker implements LogChecker {
 
     private static final String READ_LOG_FILE_OPERATION = "read-log-file";
 
@@ -28,30 +28,30 @@ public final class ModelNodeLogReader implements LogReader {
     private final boolean readLogFromEnd;
 
     /**
-     * Create an instance of log watcher. File will be read from the end (tail).
+     * Create an instance of log checker. File will be read from the end (tail).
      * @param client client which will be used to invoke log file reading operation on server
      * @param countOfLines number of lines which will be read. -1 means all lines will be read. Be careful, that log
      *                     files can get pretty big and reading whole file may lead to serious memory issues.
      */
-    public ModelNodeLogReader(final OnlineManagementClient client, final int countOfLines) {
+    public ModelNodeLogChecker(final OnlineManagementClient client, final int countOfLines) {
         this(client, countOfLines, true);
     }
 
     /**
-     * Create an instance of log watcher
+     * Create an instance of log checker
      * @param client client which will be used to invoke log file reading operation on server
      * @param countOfLines number of lines which will be read. -1 means all lines will be read. Be careful, that log
      *                     files can get pretty big and reading whole file may lead to serious memory issues.
      * @param readFromEnd true if log file should be read in tail mode
      */
-    public ModelNodeLogReader(final OnlineManagementClient client, final int countOfLines, final boolean readFromEnd) {
+    public ModelNodeLogChecker(final OnlineManagementClient client, final int countOfLines, final boolean readFromEnd) {
         this.client = client;
         this.countOfLines = countOfLines;
         this.readLogFromEnd = readFromEnd;
     }
 
     @Override
-    public boolean wasLineLogged(Pattern pattern) {
+    public boolean logMatches(Pattern pattern) {
         try {
             return anyLineMatchesPredicate(readLogFileFromManagementModel(),
                     (final String line) -> pattern.matcher(line).matches());
@@ -62,7 +62,7 @@ public final class ModelNodeLogReader implements LogReader {
     }
 
     @Override
-    public boolean wasLineLogged(String subString) {
+    public boolean logContains(String subString) {
         try {
             return anyLineMatchesPredicate(readLogFileFromManagementModel(),
                     (final String line) -> line.contains(subString));
