+++ tooling-server-configuration/src/test/java/org/jboss/eap/qe/microprofile/tooling/server/log/ModelNodeLogCheckerTestCase.java
@@ -1,4 +1,4 @@
-package org.jboss.eap.qe.microprofile.tooling.server.logwatch;
+package org.jboss.eap.qe.microprofile.tooling.server.log;
 
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
@@ -15,20 +15,20 @@ import java.util.concurrent.TimeoutException;
 import java.util.regex.Pattern;
 
 /**
- * Set of tests for {@link ModelNodeLogReader} tool.
+ * Set of tests for {@link ModelNodeLogChecker} tool.
  */
 @RunWith(Arquillian.class)
-public class ModelNodeLogReaderTestCase {
+public class ModelNodeLogCheckerTestCase {
 
     @Test
     @RunAsClient
     public void testLineMatchedPatternClient() throws ConfigurationException, IOException, TimeoutException, InterruptedException {
         try (final OnlineManagementClient client = ManagementClientProvider.onlineStandalone()) {
-            final LogReader logReader = new ModelNodeLogReader(client, 10, true);
+            final LogChecker logChecker = new ModelNodeLogChecker(client, 10, true);
 
             new Administration(client).reload();
 
-            Assert.assertTrue(logReader.wasLineLogged(Pattern.compile(".*WFLYSRV0025.*")));
+            Assert.assertTrue(logChecker.logMatches(Pattern.compile(".*WFLYSRV0025.*")));
         }
     }
 
@@ -36,9 +36,9 @@ public class ModelNodeLogReaderTestCase {
     @RunAsClient
     public void testLineNotMatchedPatternClient() throws ConfigurationException, IOException {
         try (final OnlineManagementClient client = ManagementClientProvider.onlineStandalone()) {
-            final LogReader logReader = new ModelNodeLogReader(client, 10, true);
+            final LogChecker logChecker = new ModelNodeLogChecker(client, 10, true);
 
-            Assert.assertFalse(logReader.wasLineLogged(Pattern.compile(".*Foooqux 42.*")));
+            Assert.assertFalse(logChecker.logMatches(Pattern.compile(".*Foooqux 42.*")));
         }
     }
 
@@ -46,11 +46,11 @@ public class ModelNodeLogReaderTestCase {
     @RunAsClient
     public void testLineContainedSubstringClient() throws ConfigurationException, IOException, TimeoutException, InterruptedException {
         try (final OnlineManagementClient client = ManagementClientProvider.onlineStandalone()) {
-            final LogReader logReader = new ModelNodeLogReader(client, 10, true);
+            final LogChecker logChecker = new ModelNodeLogChecker(client, 10, true);
 
             new Administration(client).reload();
 
-            Assert.assertTrue(logReader.wasLineLogged("WFLYSRV0025"));
+            Assert.assertTrue(logChecker.logContains("WFLYSRV0025"));
         }
     }
 
@@ -58,9 +58,9 @@ public class ModelNodeLogReaderTestCase {
     @RunAsClient
     public void testLineNotContainedSubstringClient() throws ConfigurationException, IOException {
         try (final OnlineManagementClient client = ManagementClientProvider.onlineStandalone()) {
-            final LogReader logReader = new ModelNodeLogReader(client, 10, true);
+            final LogChecker logChecker = new ModelNodeLogChecker(client, 10, true);
 
-            Assert.assertFalse(logReader.wasLineLogged("Foooqux 42"));
+            Assert.assertFalse(logChecker.logContains("Foooqux 42"));
         }
     }
